+++ backend/src/service/mod.rs
@@ -1,9 +1,6 @@
 pub mod beats;
 pub mod clip;
 pub mod directories;
-// pub mod download_ffmpeg;
-// pub mod ffmpeg;
-// pub mod ffprobe;
 pub mod commands;
 pub mod funscript;
 pub mod generator;